// src/app/shared/models/get-customer-info.model.ts
export interface CUSTOMERINFO {
  customerId: number;
  customerNo: number;
  customerName: string;
  mobileNumber: string;
  accounts: Account[];
  loanDetails: LoanDetail[];
}

export interface Account {
  accountType: string;
  type: string;
  accountList: AccountList[];
}

export interface AccountList {
  accountNo: string;
  accountBranch: string;
  accountCurrency: string;
  accountBalance: number;
}

export interface LoanDetail {
  accountType: string;
  type: string;
  accountList: AccountList2[];
}

export interface AccountList2 {
  accountNo: string;
  accountBranch: string;
  accountCurrency: string;
  accountBalance: number;
}